requirements.txt
falcon
prometheus-client
werkzeug
json-logging-py
gunicorn
gevent
sqlalchemy-utils
marshmallow-sqlalchemy
falcon_marshmallow
falcon-openapi
pandas
requests